requirements.txt
nextcord
dotenv-python
python-dotenv